chore: configure Metro to consume source directly
This will allow developers to modify the TypeScript code and see changes
without having to run `tsc --watch` separately.
