+++ src/WebView.ios.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {
   UIManager as NotTypedUIManager,
   View,
-  requireNativeComponent,
   NativeModules,
   Image,
   findNodeHandle,
@@ -10,6 +9,7 @@ import {
 } from 'react-native';
 import invariant from 'invariant';
 
+import RNCWebView from "./WebViewNativeComponent.ios";
 import {
   defaultOriginWhitelist,
   createOnShouldStartLoadWithRequest,
@@ -49,10 +49,6 @@ const processDecelerationRate = (
 };
 
 const RNCWebViewManager = NativeModules.RNCWebViewManager as ViewManager;
-
-const RNCWebView: typeof NativeWebViewIOS = requireNativeComponent(
-  'RNCWebView',
-);
 
 class WebView extends React.Component<IOSWebViewProps, State> {
   static defaultProps = {
@@ -202,7 +198,7 @@ class WebView extends React.Component<IOSWebViewProps, State> {
     } else {
       console.warn('Encountered an error loading page', event.nativeEvent);
     }
-    
+
     if (onLoadEnd) {
       onLoadEnd(event);
     }
