+++ src/WebView.macos.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {
   UIManager as NotTypedUIManager,
   View,
-  requireNativeComponent,
   NativeModules,
   Image,
   findNodeHandle,
@@ -10,6 +9,7 @@ import {
 } from 'react-native';
 import invariant from 'invariant';
 
+import RNCWebView from "./WebViewNativeComponent.ios";
 import {
   defaultOriginWhitelist,
   createOnShouldStartLoadWithRequest,
@@ -37,10 +37,6 @@ const UIManager = NotTypedUIManager as RNCWebViewUIManagerMacOS;
 const { resolveAssetSource } = Image;
 
 const RNCWebViewManager = NativeModules.RNCWebViewManager as ViewManager;
-
-const RNCWebView: typeof NativeWebViewMacOS = requireNativeComponent(
-  'RNCWebView',
-);
 
 class WebView extends React.Component<MacOSWebViewProps, State> {
   static defaultProps = {
@@ -189,7 +185,7 @@ class WebView extends React.Component<MacOSWebViewProps, State> {
     } else {
       console.warn('Encountered an error loading page', event.nativeEvent);
     }
-    
+
     if (onLoadEnd) {
       onLoadEnd(event);
     }
