+++ src/WebViewNativeComponent.android.ts
@@ -1,0 +1,8 @@
+import { requireNativeComponent } from "react-native";
+import type { NativeWebViewAndroid } from "./WebViewTypes";
+
+const RNCWebView: typeof NativeWebViewAndroid = requireNativeComponent(
+  'RNCWebView',
+);
+
+export default RNCWebView;
