+++ src/WebViewNativeComponent.ios.ts
@@ -1,0 +1,8 @@
+import { requireNativeComponent } from "react-native";
+import type { NativeWebViewIOS } from "./WebViewTypes";
+
+const RNCWebView: typeof NativeWebViewIOS = requireNativeComponent(
+  'RNCWebView',
+);
+
+export default RNCWebView;
