+++ src/WebViewNativeComponent.windows.ts
@@ -1,0 +1,8 @@
+import { requireNativeComponent } from "react-native";
+import type { NativeWebViewWindows } from "./WebViewTypes";
+
+const RCTWebView: typeof NativeWebViewWindows = requireNativeComponent(
+  'RCTWebView',
+);
+
+export default RCTWebView;
