+++ src/index.ts
@@ -1,0 +1,4 @@
+import WebView from './WebView';
+
+export { WebView };
+export default WebView;
